fix: 메서드의 반환 타입을 수정한다

+++ src/main/java/org/timinggame/api/room/repository/PlayerRepository.java
@@ -2,5 +2,5 @@ package org.timinggame.api.room.repository;
 
 public interface PlayerRepository {
 
-    Long getPlayerCount(final Long roomId);
+    int getPlayerCount(final Long roomId);
 }
